+++ resources/views/jobs/saved_job.blade.php
@@ -25,9 +25,13 @@
                     Updated at: {{$job->updated_at}}
                     <br>
                     <a href='/jobs/apply/{{$job->id}}'>Apply</a> |
-                    <a href='/jobs/edit/{{$job->id}}'>Edit</a> |
-                    <a href="/jobs/delete/{{$job->id}}">Delete</a> |
-                    <a href="/jobs/unsave/{{$job->id}}">Unsave</a><br>
+                    @if(Auth::check())
+                        @if(($job->author_id == $user->id)or($user->role_id > 1))
+                            <a href='/jobs/edit/{{$job->id}}'>Edit</a> |
+                            <a href="/jobs/delete/{{$job->id}}">Delete</a> |
+                        @endif
+                    @endif
+                    <a href="/jobs/unsave/{{$job->id}}">UnSave</a><br>
                 </section>
             @endforeach
         </div>
